+++ src/main/java/org/nineml/coffeefilter/InvisibleXml.java
@@ -3,8 +3,8 @@ package org.nineml.coffeefilter;
 import org.nineml.coffeefilter.exceptions.IxmlException;
 import org.nineml.coffeefilter.model.Ixml;
 import org.nineml.coffeefilter.model.IxmlContentHandler;
-import org.nineml.coffeefilter.utils.GrammarSniffer;
-import org.nineml.coffeefilter.utils.IxmlInputBuilder;
+import org.nineml.coffeefilter.util.GrammarSniffer;
+import org.nineml.coffeefilter.util.IxmlInputBuilder;
 import org.nineml.coffeegrinder.exceptions.CoffeeGrinderException;
 import org.nineml.coffeegrinder.parser.HygieneReport;
 import org.nineml.coffeegrinder.parser.SourceGrammar;
